Updating readme + refactoring
