--- rag/knowledge-bases/features-examples/02-optimizing-accuracy-retrieved-results/utils/requirements.txt
@@ -1,13 +1,0 @@
-boto3
-opensearch-py
-botocore
-awscli
-retrying
-ragas==0.1.9
-ipywidgets
-iprogress
-langchain
-langchain_aws
-langchain_community
-s3fs
-requests
