--- rag/knowledge-bases/features-examples/02-optimizing-accuracy-retrieved-results/utils/requirements_haystack.txt
@@ -1,8 +1,0 @@
-boto3
-opensearch-py
-botocore
-awscli
-retrying
-haystack-ai
-ipywidgets
-iprogress
