+++ rag/knowledge-bases/features-examples/03-advanced-concepts/reranking/README.md
@@ -17,9 +17,9 @@ In summary, reranking allows the chatbot to filter out poor responses and pick t
 ## Getting started
 To see the reranking model in action, run the following notebooks in the order for complete detail and an analysis.
 
-* [01_deploy_reranking_model_sm.ipynb](01_deploy-reranking-model-sm.ipynb) - This notebook deploys an open source reranking model [bge-reranker-large](https://huggingface.co/BAAI/bge-reranker-large) to Amazon SageMaker as a realtime endpoint. The endpoint will be used as an example for a RAG two stage retrieval application in a separate notebook. This notebook is a prerequisite for reranking model integration with Knowledge Bases for Bedrock.
+* [01_deploy_reranking_model_sm.ipynb](./01_deploy_reranking_model_sm.ipynb) - This notebook deploys an open source reranking model [bge-reranker-large](https://huggingface.co/BAAI/bge-reranker-large) to Amazon SageMaker as a realtime endpoint. The endpoint will be used as an example for a RAG two stage retrieval application in a separate notebook. This notebook is a prerequisite for reranking model integration with Knowledge Bases for Bedrock.
 
-* [02_kb-reranker.ipynb](02_kb-reranker.ipynb) - This notebook leverages a reranking model deployed in [01_deploy_reranking_model_sm.ipynb](01_deploy-reranking-model-sm.ipynb) for building a RAG application for a sample dataset. The notebook creates a knowledge bases for Bedrock with Amazon OpenSearch Serverless as the vector database, and uses the Bedrock Agent runtime API and SageMaker to orchestrate the RAG two stage retrieval process. The notebook also provides an in depth evaluation of a two stage retrieval versus a standard RAG approach using an open source RAG evaluation framework  [RAGAS](https://github.com/explodinggradients/ragas). 
+* [002_kb_reranker.ipynb](./02_kb_reranker.ipynb) - This notebook leverages a reranking model deployed in [01_deploy_reranking_model_sm.ipynb](./01_deploy_reranking_model_sm.ipynb) for building a RAG application for a sample dataset. The notebook creates a knowledge bases for Bedrock with Amazon OpenSearch Serverless as the vector database, and uses the Bedrock Agent runtime API and SageMaker to orchestrate the RAG two stage retrieval process. The notebook also provides an in depth evaluation of a two stage retrieval versus a standard RAG approach using an open source RAG evaluation framework  [RAGAS](https://github.com/explodinggradients/ragas). 
 
 ## Contributing
