+++ rag/knowledge-bases/features-examples/04-infrastructure/e2e-rag-using-bedrock-kb-cfn/templates/main-template-out.yml
@@ -1,5 +1,5 @@
 AWSTemplateFormatVersion: '2010-09-09'
-Description: end-to-end RAG solution using Knowledgebases for Amazon Bedrock
+Description: end-to-end RAG solution using Amazon Bedrock Knowledge Bases
 Parameters:
   Q01pInputBucketName:
     Type: String
