+++ rag/knowledge-bases/features-examples/requirements.txt
@@ -14,3 +14,4 @@ requests
 pypdf
 tqdm
 pandas
+haystack-ai
