+++ rag/knowledge-bases/use-case-examples/agentic-rag-using-metadata-filters/templates/2-knowledgebase-infra.yaml
@@ -1,5 +1,5 @@
 AWSTemplateFormatVersion: '2010-09-09'
-Description: Template for creating Knowledge Bases for Amazon Bedrock 
+Description: Template for creating Amazon Bedrock Knowledge Bases 
 
 Parameters:
     SolutionId:
