+++ rag/knowledge-bases/use-case-examples/rag-using-structured-unstructured-data/README.md
@@ -16,14 +16,14 @@ If you don't have these prerequisites, follow these steps:
 
 #### Prerequisites
 **Note:** This notebook assumes that you have
-1. created a knowledge base for Amazon Bedrock using unstructred data
+1. created a Amazon Bedrock Knowledge Bases using unstructred data
 2. have data available for querying via SQL in Amazon Athena.
 
 If you haven't met the prerequisite, please follow these steps:
 
 1. Create a knowledge base and ingest your documents by following this [01_create_ingest_documents_test_kb_multi_ds.ipynb](https://github.com/aws-samples/amazon-bedrock-samples/blob/main/knowledge-bases/features-examples/01-rag-concepts/01_create_ingest_documents_test_kb_multi_ds.ipynb).
 2. Note down the knowledge base ID, as you'll need it later in this notebook.
-3. If you need to use synthetic data for testing, refer to this [link](https://github.com/aws-samples/amazon-bedrock-samples/tree/main/knowledge-bases/features-examples/synthetic_dataset) to get synthetic text data that you can use to create your knowledge base for Amazon bedrock.
+3. If you need to use synthetic data for testing, refer to this [link](https://github.com/aws-samples/amazon-bedrock-samples/tree/main/knowledge-bases/features-examples/synthetic_dataset) to get synthetic text data that you can use to create your Amazon Bedrock Knowledge Bases.
 4. To create synthetic structured data, you can run [0-create-dummy-structured-data.ipynb](https://github.com/aws-samples/amazon-bedrock-samples/tree/main/knowledge-bases/use-case-examples/rag-using-structured-unstructured-data/0-create-dummy-structured-data.ipynb) notebook and then use [1_create_sql_dataset_optional.ipynb](https://github.com/aws-samples/amazon-bedrock-samples/tree/main/knowledge-bases/use-case-examples/rag-using-structured-unstructured-data/1_create_sql_dataset_optional.ipynb) notebook to create a database and table in Amazon Athena.
 
 **Note**: The `custom_database_retriever.py` file currently uses table schema for a retail order website generated by using [0-create-dummy-structured-data.ipynb](https://github.com/aws-samples/amazon-bedrock-samples/tree/main/knowledge-bases/use-case-examples/rag-using-structured-unstructured-data/0-create-dummy-structured-data.ipynb) notebook. If you choose to use a different dataset, please update the schema for tables and table information inside `custom_database_retriever.py` file.
